requirements.txt
streamlit>=1.33.0
groq>=0.5.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-huggingface>=0.0.1
sentence-transformers>=2.2.0
chromadb>=0.4.0
pypdf>=4.0.0
fpdf2>=2.7.0
python-dotenv>=1.0.0
pillow>=10.0.0
reportlab>=4.1.0
pysqlite3-binary
